Time: 15 ms (78.26%) | Memory: 19.6 MB (21.00%) - LeetSync

+++ 283-move-zeroes/move-zeroes.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     void moveZeroes(vector<int>& nums) {
-        int zc=0,p=0,s=nums.size();
+     short int zc=0,p=0,s=nums.size();
         for(int i=0;i<s;i++)
         {
             if(nums[i]==0)
